+++ access.log
@@ -253,3 +253,10 @@
 ::1 - - [26/Jan/2022:00:13:07 +0000] "PATCH /api/contacts/61f08de0690f83a1008e97a3/favorite HTTP/1.1" 200 250 "-" "PostmanRuntime/7.28.4"
 ::1 - - [26/Jan/2022:00:19:34 +0000] "DELETE /api/contacts/61f08de0690f83a1008e97a3 HTTP/1.1" 200 29 "-" "PostmanRuntime/7.28.4"
 ::1 - - [26/Jan/2022:00:20:16 +0000] "GET /api/contacts HTTP/1.1" 200 467 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:01:07:11 +0000] "POST /api/users/signup HTTP/1.1" 201 62 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:01:14:24 +0000] "POST /api/users/verify HTTP/1.1" 200 37 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:01:15:48 +0000] "GET /api/users/verify/BzqVqVGIIjnP431xj9brw HTTP/1.1" 200 37 "-" "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/97.0.4692.71 Safari/537.36"
+::1 - - [26/Jan/2022:01:19:41 +0000] "POST /api/users/login HTTP/1.1" 200 244 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:01:21:54 +0000] "GET /api/users/current HTTP/1.1" 200 62 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:01:23:22 +0000] "PATCH /api/users/avatars HTTP/1.1" 200 53 "-" "PostmanRuntime/7.28.4"
+::1 - - [26/Jan/2022:01:25:12 +0000] "GET /api/users/logout HTTP/1.1" 204 - "-" "PostmanRuntime/7.28.4"
