+++ controllers/users/getCurrent.js
@@ -1,0 +1,10 @@
+const getCurrent = async (req, res) => {
+    const { subscription, email } = req.user;
+    res.json({
+        user: {
+            subscription,
+            email
+        }
+    })
+}
+module.exports = getCurrent;
