+++ controllers/users/index.js
@@ -1,2 +1,8 @@
 const signup = require("./signup");
-module.exports = {signup};
+const avatars = require("./avatars");
+const login = require("./login");
+const logout = require("./logout");
+const getCurrent = require("./getCurrent");
+const verify = require("./verify");
+const getVerifyToken = require("./getVerifyToken");
+module.exports = {signup, avatars, login, logout, getCurrent, verify, getVerifyToken};
